feat: Add provider policy, results, and serialization modules
- Introduced `provider_policy.py` to manage market data provider configurations and warnings.
- Created `provider_results.py` to define data structures for snapshot and refresh results, including paths and counts.
- Implemented `provider_serialization.py` for handling serialization of provider payloads and metadata.
- Updated tests to include new pipeline facade exports for the market data module.
